Fix sens level, add tests

+++ src/fft.py
@@ -32,7 +32,7 @@ class FFT:
 
     @staticmethod
     def _change_spectrum_level(spectrum: np.array,
-                               level: int = FFT_DEFAULT_LEVEL_FACTOR) -> np.array:
+                               level: float = FFT_DEFAULT_LEVEL_FACTOR) -> np.array:
         if level == 1:
             return spectrum
         return spectrum * level
@@ -40,7 +40,7 @@ class FFT:
     @staticmethod
     def data_ready_to_plot(raw_data: str,
                            decimation_factor: int,
-                           sensitivity_level: int) -> np.array:
+                           sensitivity_level: float) -> np.array:
         data = FFT._get_time_data(raw_data)
         data = FFT._get_amplitude_spectrum(data)
         data = FFT._decimate_spectrum(data, decimation_factor)
